Error updates for the TVM calculator
